Refactor Projects section to use ContainerProject component
The Projects section in src/sections/Projects.tsx has been refactored to use the newly imported ContainerProject component. This change helps to maintain consistency and reusability of components across the App. The ContainerProject component already contains the required className so it has improved the readability of the code by replacing repeated CSS classes.